+++ src/main/java/com/besson/arknights/painting/ModPaintingVariants.java
@@ -60,6 +60,8 @@ public class ModPaintingVariants {
     public static final PaintingVariant STREAM_OF_CONSCIOUSNESS_PAINTING = registerPaintingVariant("stream_of_consciousness_painting", new PaintingVariant(16, 16));
     public static final PaintingVariant THE_GOLDEN_ERA = registerPaintingVariant("the_golden_era", new PaintingVariant(16, 16));
     public static final PaintingVariant VOYAGE_IMPRESSION_KIDS_PAINTING = registerPaintingVariant("voyage_impression_kids_painting", new PaintingVariant(16, 16));
+    public static final PaintingVariant BAND_PROMOTIONAL_POSTER = registerPaintingVariant("band_promotional_poster", new PaintingVariant(16, 32));
+    public static final PaintingVariant SARKAZ_ROCK_N_ROLL_POSTER = registerPaintingVariant("sarkaz_rock_n_roll_poster", new PaintingVariant(16, 32));
 
     private static PaintingVariant registerPaintingVariant(String id, PaintingVariant variant) {
         return Registry.register(Registries.PAINTING_VARIANT, new Identifier(ArknightsFurniture.MOD_ID, id), variant);